ConcurrentCollections.jl

ConcurrentCollections.ConcurrentDict

ConcurrentCollections.Delete

ConcurrentCollections.ConcurrentQueue

ConcurrentCollections.ConcurrentStack

ConcurrentCollections.WorkStealingDeque

ConcurrentCollections.Keep

ConcurrentCollections.modify!

ConcurrentCollections.tryget

ConcurrentCollections.trypop!

ConcurrentCollections.trypopfirst!

Collections

ConcurrentCollections.ConcurrentQueue — Type

ConcurrentQueue{T}()

Concurrent queue of objects of type T.

Use push! to insert an element at the tail and trypopfirst! to retrieve and remove an element at the head.

Implementation detail: It implements the Michael and Scott queue.

Examples

julia> using ConcurrentCollections

julia> queue = ConcurrentQueue{Int}();

julia> push!(queue, 1);

julia> push!(queue, 2);

julia> popfirst!(queue)
1

julia> trypopfirst!(queue)
Some(2)

julia> trypopfirst!(queue)  # returns nothing

ConcurrentCollections.ConcurrentStack — Type

ConcurrentStack{T}()

Concurrent stack of objects of type T.

Use push! to insert an element and trypop! to retrieve and remove an element.

It implements the Treiber stack.

Examples

julia> using ConcurrentCollections

julia> stack = ConcurrentStack{Int}();

julia> push!(stack, 1);

julia> push!(stack, 2);

julia> pop!(stack)
2

julia> trypop!(stack)
Some(1)

julia> trypop!(stack)  # returns nothing

ConcurrentCollections.WorkStealingDeque — Type

WorkStealingDeque{T}()

Concurrent work-stealing "deque" of objects of type T.

This is not a full deque in the sense that:

push! and trypop! operating at the tail of the collection can only be executed by a single task.

trypopfirst! (aka steal) for retrieving and removing an element at the head can be invoked from any tasks. However, there is no pushfirst!.

Implementation detail: It implements the dynamic circular work-stealing deque by Chase and Lev (2005).

Examples

julia> using ConcurrentCollections

julia> deque = WorkStealingDeque{Int}();

julia> push!(deque, 1);

julia> push!(deque, 2);

julia> push!(deque, 3);

julia> trypop!(deque)
Some(3)

julia> fetch(Threads.@spawn trypopfirst!(deque))
Some(1)

julia> fetch(Threads.@spawn popfirst!(deque))
2

julia> trypopfirst!(deque)  # returns nothing

Functions

ConcurrentCollections.trypop! — Function

trypop!(collection) -> Some(value::T) or nothing

Try to pop a value from the tail of collection. Return Some(value) if it is non-empty. Return nothing if empty.

Examples

julia> using ConcurrentCollections

julia> stack = ConcurrentStack{Int}();

julia> push!(stack, 1);

julia> trypop!(stack)
Some(1)

julia> trypop!(stack)  # returns nothing

ConcurrentCollections.trypopfirst! — Function

trypopfirst!(collection) -> Some(value::T) or nothing

Try to pop a value from the head of collection. Return Some(value) if it is non-empty. Return nothing if empty.

Examples

julia> using ConcurrentCollections

julia> queue = ConcurrentQueue{Int}();

julia> push!(queue, 1);

julia> trypopfirst!(queue)
Some(1)

julia> trypopfirst!(queue)  # returns nothing

Experimental

ConcurrentCollections.ConcurrentDict — Type

ConcurrentDict{K,V}()

Concurrent dictionary. All operations are lock-free except when the dictionary is resized.

Warning

ConcurrentDict is experimental because it cannot be implemented with the builtin atomics.

Note

Although tasks wait on concurrent modifications (e.g., setindex!) during resize, the worker threads participate in the resize to avoid wasting CPU resources.

Examples

julia> using ConcurrentCollections

julia> dict = ConcurrentDict{String,Int}();

julia> dict["hello"] = 1;

julia> dict["hello"]
1

ConcurrentCollections.modify! — Function

modify!(f, dict::ConcurrentDict{K,V}, key::K) -> y

Atomically update key slot of dict using a function f.

If key does not exist, f is called with nothing. The call f(nothing) must return either (1) nothing to keep the slot unoccupied or (2) Some(value::V) to insert value.

If key exist, f is called with a ref such that ref[] retrieves the value corresponding to the key. The call f(ref) must return either (1) nothing to delete the slot, (2) Some(value′::V) to insert value, (3) Keep(ans) to return y = Keep(ans) from modify!, or (4) Delete(ans) to delete slot and return a value y = Delete(ans) from modify!.

The function f may be called more than once if multiple tasks try to modify the dictionary.

Examples

julia> using ConcurrentCollections

julia> dict = ConcurrentDict{String,Int}();

julia> modify!(dict, "hello") do _
           Some(1)
       end
Some(1)

julia> modify!(dict, "hello") do ref
           Some(something(ref[]) + 1)
       end
Some(2)

ConcurrentCollections.tryget — Function

tryget(dict::ConcurrentDict{K,V}, key) -> Some(value::T) or nothing

ConcurrentCollections.Keep — Type

Keep(ans)

A special type used in modify! to indicate that a slot should be remain unchanged while propagating the result ans of some computation to the caller.

That is to say,

y = modify!(dict, key) do value
    Keep(f(something(value)))
end
y[]

is an optimization of

r = Ref{Any}()
modify!(dict, key) do value
    r[] = f(something(value))
    Some(value)
end
r[]

ConcurrentCollections.Delete — Type

Delete(ans)

A special type used in modify! to indicate that a slot should be removed.

That is to say

y = modify!(dict, key) do value
    Delete(f(something(value)))
end
y[]

is an optimization of

r = Ref{Any}()
modify!(dict, key) do value
    r[] = f(something(value))
    nothing
end
r[]